# backend/docs/swagger.yaml
basePath: /
definitions:
...
  domain.AppType:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    - 8
    - 9
    - 10
    - 11
    format: int32
    type: integer
    x-enum-varnames:
    - AppTypeWeb
    - AppTypeWidget
    - AppTypeDingTalkBot
    - AppTypeFeishuBot
    - AppTypeWechatBot
    - AppTypeWechatServiceBot
    - AppTypeDisCordBot
    - AppTypeWechatOfficialAccount
    - AppTypeOpenAIAPI
    - AppTypeWecomAIBot
    - AppTypeLarkBot
  domain.AuthUserInfo:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      username:
        type: string
    type: object
...
  domain.BrowserCount:
    properties:
      count:
        type: integer
      name:
        type: string
    type: object
...
  domain.ConversationDistribution:
    properties:
      app_id:
        type: string
      app_type:
        $ref: '#/definitions/domain.AppType'
      count:
        type: integer
    type: object
...
  domain.HotBrowser:
    properties:
      browser:
        items:
          $ref: '#/definitions/domain.BrowserCount'
        type: array
      os:
        items:
          $ref: '#/definitions/domain.BrowserCount'
        type: array
    type: object
  domain.HotPage:
    properties:
      count:
        type: integer
      node_id:
        type: string
      node_name:
        type: string
      scene:
        $ref: '#/definitions/domain.StatPageScene'
    type: object
  domain.HotRefererHost:
    properties:
      count:
        type: integer
      referer_host:
        type: string
    type: object
  domain.IPAddress:
    properties:
      city:
        type: string
      country:
        type: string
      ip:
        type: string
      province:
        type: string
    type: object
...
  domain.InstantCountResp:
    properties:
      count:
        type: integer
      time:
        type: string
    type: object
  domain.InstantPageResp:
    properties:
      created_at:
        type: string
      info:
        $ref: '#/definitions/domain.AuthUserInfo'
      ip:
        type: string
      ip_address:
        $ref: '#/definitions/domain.IPAddress'
      node_id:
        type: string
      node_name:
        type: string
      scene:
        $ref: '#/definitions/domain.StatPageScene'
      user_id:
        type: integer
    type: object
...
  domain.PWResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  domain.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  domain.StatPageScene:
    enum:
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - StatPageSceneWelcome
    - StatPageSceneNodeDetail
    - StatPageSceneChat
    - StatPageSceneLogin
...
  v1.StatCountResp:
    properties:
      conversation_count:
        type: integer
      ip_count:
        type: integer
      page_visit_count:
        type: integer
      session_count:
        type: integer
    type: object
...
info:
  contact: {}
  description: panda-wiki API documentation
  title: panda-wiki API
  version: "1.0"
paths:
...
  /api/v1/stat/browsers:
    get:
      consumes:
      - application/json
      description: 客户端统计
      parameters:
      - enum:
        - 1
        - 7
        - 30
        - 90
        in: query
        name: day
        type: integer
        x-enum-varnames:
        - StatDay1
        - StatDay7
        - StatDay30
        - StatDay90
      - in: query
        name: kb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.HotBrowser'
              type: object
      security:
      - bearerAuth: []
      summary: 客户端统计
      tags:
      - stat
  /api/v1/stat/conversation_distribution:
    get:
      consumes:
      - application/json
      description: 问答来源
      parameters:
      - enum:
        - 1
        - 7
        - 30
        - 90
        in: query
        name: day
        type: integer
        x-enum-varnames:
        - StatDay1
        - StatDay7
        - StatDay30
        - StatDay90
      - in: query
        name: kb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.ConversationDistribution'
                  type: array
              type: object
      security:
      - bearerAuth: []
      summary: 问答来源
      tags:
      - stat
  /api/v1/stat/count:
    get:
      consumes:
      - application/json
      description: 全局统计
      parameters:
      - enum:
        - 1
        - 7
        - 30
        - 90
        in: query
        name: day
        type: integer
        x-enum-varnames:
        - StatDay1
        - StatDay7
        - StatDay30
        - StatDay90
      - in: query
        name: kb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.StatCountResp'
              type: object
      security:
      - bearerAuth: []
      summary: 全局统计
      tags:
      - stat
  /api/v1/stat/geo_count:
    get:
      consumes:
      - application/json
      description: 用户地理分布
      parameters:
      - enum:
        - 1
        - 7
        - 30
        - 90
        in: query
        name: day
        type: integer
        x-enum-varnames:
        - StatDay1
        - StatDay7
        - StatDay30
        - StatDay90
      - in: query
        name: kb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - bearerAuth: []
      summary: 用户地理分布
      tags:
      - stat
  /api/v1/stat/hot_pages:
    get:
      consumes:
      - application/json
      description: 热门文档
      parameters:
      - enum:
        - 1
        - 7
        - 30
        - 90
        in: query
        name: day
        type: integer
        x-enum-varnames:
        - StatDay1
        - StatDay7
        - StatDay30
        - StatDay90
      - in: query
        name: kb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.HotPage'
                  type: array
              type: object
      security:
      - bearerAuth: []
      summary: 热门文档
      tags:
      - stat
  /api/v1/stat/instant_count:
    get:
      consumes:
      - application/json
      description: GetInstantCount
      parameters:
      - in: query
        name: kb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.InstantCountResp'
                  type: array
              type: object
      security:
      - bearerAuth: []
      summary: GetInstantCount
      tags:
      - stat
  /api/v1/stat/instant_pages:
    get:
      consumes:
      - application/json
      description: GetInstantPages
      parameters:
      - in: query
        name: kb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.InstantPageResp'
                  type: array
              type: object
      security:
      - bearerAuth: []
      summary: GetInstantPages
      tags:
      - stat
  /api/v1/stat/referer_hosts:
    get:
      consumes:
      - application/json
      description: 来源域名
      parameters:
      - enum:
        - 1
        - 7
        - 30
        - 90
        in: query
        name: day
        type: integer
        x-enum-varnames:
        - StatDay1
        - StatDay7
        - StatDay30
        - StatDay90
      - in: query
        name: kb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.HotRefererHost'
                  type: array
              type: object
      security:
      - bearerAuth: []
      summary: 来源域名
      tags:
      - stat
...
securityDefinitions:
  bearerAuth:
    description: Type "Bearer" + a space + your token to authorize
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"